procurement_service/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Cors==3.0.10
SQLAlchemy==1.4.23
python-dotenv==0.19.0
requests==2.26.0
graphene==2.1.9
graphene-sqlalchemy==2.3.0
flask-graphql==2.0.1 